Rename class for api.
