+++ lib/mongo_browser.rb
@@ -28,4 +28,4 @@ require "mongo_browser/models/collection"
 require "mongo_browser/models/document"
 
 require "mongo_browser/application"
-require "mongo_browser/application/api"
+require "mongo_browser/api"
